+++ app/src/main/java/net/programmierecke/radiodroid2/FragmentStarred.java
@@ -85,10 +85,16 @@ public class FragmentStarred extends Fragment implements IAdapterRefreshable, Ob
         SharedPreferences sharedPref = PreferenceManager.getDefaultSharedPreferences(getContext());
         boolean useIconView = sharedPref.getBoolean("icons_only_favorites_style", false);
         boolean loadIcons = sharedPref.getBoolean("load_icons", true);
+        boolean isAndroidAuto = Utils.isAndroidAutoMode(getContext());
+        
+        // Force icon matrix display in Android Auto mode - disable list option
+        if (isAndroidAuto) {
+            useIconView = true;
+        }
         
         if (BuildConfig.DEBUG) {
             Log.d(TAG, "FragmentStarred: useIconView=" + useIconView + ", loadIcons=" + loadIcons + 
-                ", isAndroidAuto=" + Utils.isAndroidAutoMode(getContext()));
+                ", isAndroidAuto=" + isAndroidAuto + (isAndroidAuto ? " (forced icon view)" : ""));
         }
         
         if (useIconView) {